♻️ refactor : 텍스트 색상 color picker 적용 #201
